+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added `keyple-plugin-storage-card-java-api:1.0.0` dependency to enable storage card support. 
 
 ## [3.0.0] - 2025-02-10
 ### Breaking changes